Move testutil to internal, copied in contrib as well
